Modularized Public Map View
